[Refactor] - Moves missing files from info to membership module
