+++ membership/src/main/res/values-es/strings.xml
@@ -7,4 +7,5 @@
     <string name="switchCardHolder">Cambiar titular de tarjeta</string>
     <string name="memberCard">Tarjeta del Miembro</string>
     <string name="accessMemberCardLabel">Tarjeta de acceso para miembros</string>
+    <string name="expires">Vence %s</string>
 </resources>
