+++ membership/src/main/res/values-zh/strings.xml
@@ -7,4 +7,5 @@
     <string name="switchCardHolder">交换持卡人</string>
     <string name="memberCard">会员卡</string>
     <string name="accessMemberCardLabel">访问会员卡</string>
+    <string name="expires">截止 %s</string>
 </resources>
