+++ membership/src/main/res/values/strings.xml
@@ -12,4 +12,5 @@
     <string name="errorDialogTitle">Error</string>
     <string name="ok" translatable="false">@android:string/ok</string>
     <string name="accessMemberCardLabel">Access Member Card</string>
+    <string name="expires">Expires %s</string>
 </resources>
